SoundSettings: Suppress a race condition crash
mNotificationPreference is attempting to refresh itself using the setEnabled
function before the recyclerview is finished computing the layout which causes the
fragment to crash with an IllegalStateException. Suppress this crash to let the fragment
load.

I have not found a definite pattern to this yet, but it does happen on rare occasions

Change-Id: Iaadd9a7069c8bbf3a7ea2ef6392856799b54d9af